+++ src/extension/connector/components/DappPlateHeader.vue
