+++ src/extension/connector/views/AuthView.vue
@@ -5,7 +5,6 @@ import { helpers, requiredUnless } from "@vuelidate/validators";
 import { useEventListener } from "@vueuse/core";
 import { AlertCircleIcon } from "lucide-vue-next";
 import { useWalletStore } from "@/stores/walletStore";
-import DappPlateHeader from "@/components/DappPlateHeader.vue";
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
@@ -15,6 +14,7 @@ import { useToast } from "@/components/ui/toast";
 import { signAuthMessage } from "@/chains/ergo/signing";
 import { PasswordError } from "@/common/errors";
 import { connectedDAppsDbService } from "@/database/connectedDAppsDbService";
+import DappPlateHeader from "@/extension/connector/components/DappPlateHeader.vue";
 import { AsyncRequest } from "@/extension/connector/rpc/asyncRequestQueue";
 import { error, InternalRequest, success } from "@/extension/connector/rpc/protocol";
 import { queue } from "@/extension/connector/rpc/uiRpcHandlers";
