+++ src/extension/connector/views/DataSignView.vue
@@ -9,7 +9,6 @@ import { AlertCircleIcon } from "lucide-vue-next";
 import type { JsonObject } from "type-fest";
 import { useAppStore } from "@/stores/appStore";
 import { useWalletStore } from "@/stores/walletStore";
-import DappPlateHeader from "@/components/DappPlateHeader.vue";
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
@@ -20,6 +19,7 @@ import { useToast } from "@/components/ui/toast";
 import { signMessage } from "@/chains/ergo/signing";
 import { PasswordError } from "@/common/errors";
 import { connectedDAppsDbService } from "@/database/connectedDAppsDbService";
+import DappPlateHeader from "@/extension/connector/components/DappPlateHeader.vue";
 import { AsyncRequest } from "@/extension/connector/rpc/asyncRequestQueue";
 import { error, InternalRequest, success } from "@/extension/connector/rpc/protocol";
 import { queue } from "@/extension/connector/rpc/uiRpcHandlers";
