+++ src/extension/connector/views/TransactionSignView.vue
@@ -4,9 +4,9 @@ import { SignedInput, SignedTransaction, some } from "@fleet-sdk/common";
 import { useEventListener } from "@vueuse/core";
 import { useAssetsStore } from "@/stores/assetsStore";
 import { useWalletStore } from "@/stores/walletStore";
-import DappPlateHeader from "@/components/DappPlateHeader.vue";
 import { TransactionSign } from "@/components/transaction";
 import { connectedDAppsDbService } from "@/database/connectedDAppsDbService";
+import DappPlateHeader from "@/extension/connector/components/DappPlateHeader.vue";
 import { AsyncRequest } from "@/extension/connector/rpc/asyncRequestQueue";
 import { error, InternalRequest, success } from "@/extension/connector/rpc/protocol";
 import { queue } from "@/extension/connector/rpc/uiRpcHandlers";
